Rename kms/internal/plugins/_mock -> kms/_mock
